# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para operaciones del sistema SistaL.",
        "title": "MS Operations API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8083",
    "basePath": "/api/v1",
    "paths": {
...
        "/facturas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Obtiene lista paginada de facturas con filtros opcionales",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facturas"
                ],
                "summary": "Obtener facturas",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Número de página",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Elementos por página",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de estado",
                        "name": "id_estado",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Filtrar por ID de empresa",
                        "name": "id_empresa",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha desde (YYYY-MM-DD)",
                        "name": "fecha_desde",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Fecha hasta (YYYY-MM-DD)",
                        "name": "fecha_hasta",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Búsqueda en número de factura",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de facturas con paginación",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Crea una nueva factura en el sistema",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Facturas"
                ],
                "summary": "Crear factura",
                "parameters": [
                    {
                        "description": "Datos de la factura",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateFacturaDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Factura creada exitosamente",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Datos de entrada inválidos",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error del servidor",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateFacturaDTO": {
            "type": "object",
            "required": [
                "id_empresa",
                "monto_total",
                "numero_factura"
            ],
            "properties": {
                "fecha_emision": {
                    "type": "string"
                },
                "id_empresa": {
                    "type": "integer"
                },
                "id_estado": {
                    "type": "integer"
                },
                "monto_total": {
                    "type": "number",
                    "minimum": 0
                },
                "numero_factura": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Token de autenticación JWT en formato Bearer",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}